1. Point – Changed "DOWN" to "DROP"
2. Deleted "Tetris.cpp", instead created "main.cpp"

3. Main:
a. Draws the board and the score table
b. Creates the Blocks (Left and Right)
c. The flow beings where at first it moves the block with the "true" statement so it wont ruin the border and then the same algorithm, didn't change it much except for the variables "LEFTP1" etc..
d. At the end of the if moves both blocks down and sleep of 1 second.

4. Block:
a. Added "char side" to the variables and to the constructor
b. Added "char side" to "buildBlockPoints", if the side is 'R' it adds "RIVAL_MIN_X"
c. Fixed the typo of the func "genetrate.." to "generate.."
d. Changed "moveBlock" - draw whitespace -> move -> draw symbol, in case the draw is the first draw it doesn't draw a whitespace so it won't ruin the border.

5. Created Board:
a. drawBordersAndScore calls 3 different functions (for clarity) to draw all of the needed borders including the score table

6. GameConfig:
a. Changed eKeys LEFT & RIGHT to LEFTP1 and RIGHTP1
b. Added LEFTP2 and RIGHTP2
c. Added another static constexpr int named "RIVAL_MIN_X" in order to separate the right player from the left player (adds 32 to the "MIN_X")